lex: Make lexer_force_int() handle full int64_t range.
This requires a test update because 18446744073709551615 is now in the
range accepted by lexer_force_int().

+++ include/ovn/lex.h
@@ -136,8 +136,8 @@ bool lexer_match(struct lexer *, enum lex_type);
 bool lexer_force_match(struct lexer *, enum lex_type);
 bool lexer_match_id(struct lexer *, const char *id);
 bool lexer_is_int(const struct lexer *);
-bool lexer_get_int(struct lexer *, int *value);
-bool lexer_force_int(struct lexer *, int *value);
+bool lexer_get_int(struct lexer *, int64_t *value);
+bool lexer_force_int(struct lexer *, int64_t *value);
 
 void lexer_force_end(struct lexer *);
